Funcion para editar correos
se le permite al usuario editar los correos de la bandeja de salida

+++ js/correo.js
@@ -1,8 +1,9 @@
+var conti=0;
+var conta=0;
 var LOGIN=LOGIN||
-{
+{ 
 		//funcion para guardar un usuario
-		guardarUser: function() {
-
+		guardarUser:function() {
 			var email = document.getElementById("email").value;
 			var contrasena = document.getElementById("contrasena").value;
 			debugger;
@@ -83,9 +84,9 @@ var LOGIN=LOGIN||
 },mostrar:function()
 {
   debugger;
-	document.getElementById('oculto').style.display = 'block';
-	document.getElementById('btnsesion').style.display = 'none';
-	document.getElementById('btnregistro').style.display = 'none';
+  document.getElementById('oculto').style.display = 'block';
+  document.getElementById('btnsesion').style.display = 'none';
+  document.getElementById('btnregistro').style.display = 'none';
 
 	//funcion para validar que el usuario llene todos los campos exigidos
 },validarcampos:function(){
@@ -191,35 +192,34 @@ var LOGIN=LOGIN||
   user = JSON.parse(localStorage.getItem("Online")).online;
   var cont=0;
   if(bandeborrador!=null){
-  for (var i = 0; i < bandeborrador.length; i++) { 
-    var c = bandeborrador[i];  
-    if(c!=null){
-    if(bandeborrador[i].User===user){
-      cont=cont+1;
-      if(bandeborrador.length!=null){
-        user_html = user_html + "<div onClick=' LOGIN.contoculto("+i+");' id="+i+" class=' mostrar nave panel panel-default imagenConPieDeTexto sombra row-fluid' >"+"<span class= 'glyphicon glyphicon-envelope'>"+"</span>"+"<span id='corrasun' class='text' style='color:black' >"+c.asunto+"&nbsp;"+"</span>"+
-       "<span id='corrcont' class='text' style='color:gray' maxlength='10'>" +"-"+ c.contenido+"</span>"+"<span id='trash' class='glyphicon glyphicon-trash' style='float:right' onClick='LOGIN.eliminarcorreoborrador("+i+");'></span>"+"<nav><div id="+"C"+i+"></div></nav>"+
-       "<div id="+"M"+i+" class='ocultar 'class='pr colocar'><div ><div class='animated fadeIn colornuevo sombra2'>"+
-       "<header class=he><div><p class=txt izq>"+c.asunto+"</p><div class=>"+
-       "<a title=Eliminar correo>"+
-       "<img class='padding' title='Responder mensaje' src='Imagenes/responder.png'><a/>"+
-       "<a onclick='LOGIN.editarcorreoborrador("+i+");' title=Eliminar correo><img class='padding' title='Editar mensaje' src='Imagenes/edit.png'>"+
-       "<a/><a href='eliminar.html' title='Eliminar correo'><img class='padding' src='Imagenes/trash.png'><a/>"
-       +"</div></div></header><hr>"+ "<div class='div'><p class='txt-izq'>"+c.para+"</p>"+"<div>"+"<header>" + c.asunto +"</header>"+"</div><div>"+"<p class='contenido'>"+c.contenido+"</p>"+"</div>"+"</div>"
-       +"</div></div></div>"+"</div>"
-     }else{
-      user_html = user_html + "<nav class='bandeja'>"+"<p class='hoy salto'>"+"Aca se mostraran tus mensajes que se encuentran en borrador"+"</p>"+"</nav>";
-    }
-    $('#correos_borrados').html(user_html);
+    for (var i = 0; i < bandeborrador.length; i++) { 
+      var c = bandeborrador[i];  
+      if(c!=null){
+        if(bandeborrador[i].User===user){
+          cont=cont+1;
+          if(bandeborrador.length!=null){
+            user_html = user_html + "<div id="+i+" class=' mostrar nave panel panel-default imagenConPieDeTexto sombra row-fluid' >"+"<div onClick=' LOGIN.contoculto("+i+");'><span class= 'glyphicon glyphicon-envelope'>"+"</span>"+"<span id='corrasun' class='text' style='color:black' >"+c.asunto+"&nbsp;"+"</span>"+
+            "<span id='corrcont' class='text' style='color:gray' maxlength='10'>" +"-"+ c.contenido+"</span>"+"<span id='trash' class='glyphicon glyphicon-trash' style='float:right' onClick='LOGIN.eliminarcorreoborrador("+i+");'></span></div>"+"<nav><div id="+"C"+i+"></div></nav>"+
+            "<div id="+"M"+i+" class='ocultar 'class='pr colocar'><div ><div class='animated fadeIn colornuevo sombra2'>"+
+            "<header class=he><div><p class=txt izq>"+c.asunto+"</p><div class=>"+
+            "<a title=Eliminar correo>"+
+            "<img class='padding' title='Responder mensaje' src='Imagenes/responder.png'><a/>"+
+            "<a onclick='LOGIN.editarcorreoborrador("+i+");' title=Eliminar correo><img class='padding' title='Editar mensaje' src='Imagenes/edit.png'>"+
+            "<a/><a href='eliminar.html' title='Eliminar correo'><img class='padding' src='Imagenes/trash.png'><a/>"
+            +"</div></div></header><hr>"+ "<div class='div'><p class='txt-izq'>"+c.para+"</p>"+"<div>"+"<header>" + c.asunto +"</header>"+"</div><div>"+"<p class='contenido'>"+c.contenido+"</p>"+"</div>"+"</div>"
+            +"</div></div></div>"+"</div>"
+          }else{
+            user_html = user_html + "<nav class='bandeja'>"+"<p class='hoy salto'>"+"Aca se mostraran tus mensajes que se encuentran en borrador"+"</p>"+"</nav>";
+          }
+          $('#correos_borrados').html(user_html);
 
+        }
+      }else{
+        user_html = user_html + "<nav class='bandeja'>"+"<p class='hoy salto'>"+"Aca se mostraran tus mensajes que se encuentran en borrador"+"</p>"+"</nav>";
+      }
+    }
   }
-}else{
-  user_html = user_html + "<nav class='bandeja'>"+"<p class='hoy salto'>"+"Aca se mostraran tus mensajes que se encuentran en borrador"+"</p>"+"</nav>";
-}
-}
-user_html = user_html + "<nav class='bandeja'>"+"<p class='hoy salto'>"+"Aca se mostraran tus mensajes que se encuentran en borrador"+"</p>"+"</nav>";
-}
-$('#icorreos').html(cont);
+  $('#icorreos').html(cont);
 
 },fecha:function(){
   debugger;
@@ -237,23 +237,25 @@ $('#icorreos').html(cont);
  user = JSON.parse(localStorage.getItem("Online")).online;
  var contador=0;
  if(bandenviados!=null){
- for (var i = 0; i < bandenviados.length; i++) { 
-  var e = bandenviados[i];  
-  if(bandenviados[i].User===user){
-    var contador=contador+1;
-    if(bandenviados.length!=null){
-       user_html = user_html + "<div onClick=' LOGIN.contoculto("+i+");' id="+i+" class=' mostrar nave panel panel-default imagenConPieDeTexto sombra row-fluid' >"+"<span class= 'glyphicon glyphicon-envelope'>"+"</span>"+"<span id='corrasun' class='text' style='color:black' >"+e.asunto+"&nbsp;"+"</span>"+
-       "<span id='corrcont' class='text' style='color:gray' maxlength='10'>" +"-"+ e.contenido+"</span>"+"<span id='trash' class='glyphicon glyphicon-trash' style='float:right' onClick='LOGIN.eliminarcorreoenviados("+i+");'></span>"+"<nav><div id="+"E"+i+"></div></nav>"+
-       "<div id="+"M"+i+" class='ocultar'class='pr colocar'><div ><div class= 'colornuevo class='animated fadeIn sombra2'>"+
-       "<header class=he><div><p class=txt izq>"+e.asunto+"</p><div class=>"+
-       "<a title=Eliminar correo>"+
-       "<img class='padding' title='Responder mensaje' src='Imagenes/responder.png'><a/>"+
-       "<a href='editar.html' title=Eliminar correo><img class='padding' title='Editar mensaje' src='Imagenes/edit.png'>"+
-       "<a/><a href='eliminar.html' title='Eliminar correo'><img class='padding' src='Imagenes/trash.png'><a/>"
-       +"</div></div></header><hr>"+ "<div class='div'><p class='txt-izq'>"+e.para+"</p>"+"<div>"+"<header>" + e.asunto +"</header>"+"</div><div>"+"<p class='contenido'>"+e.contenido+"</p>"+"</div>"+"</div>"
-       +"</div></div></div>"+"</div>"
-    }else{
-      user_html = user_html + "<nav class='bandeja'>"+"<p class='hoy salto'>"+"Aca se mostraran tus mensajes que se encuentran en borrador"+"</p>"+"</nav>";
+   for (var i = 0; i < bandenviados.length; i++) { 
+    var e = bandenviados[i]; 
+    if(e!=null){ 
+      if(bandenviados[i].User===user){
+        var contador=contador+1;
+        if(bandenviados.length!=null){
+         user_html = user_html + "<div onClick=' LOGIN.contoculto("+i+");' id="+i+" class=' mostrar nave panel panel-default imagenConPieDeTexto sombra row-fluid' >"+"<span class='glyphicon glyphicon-envelope'>"+"</span>"+"<span id='corrasun' class='text' style='color:black' >"+e.asunto+"&nbsp;"+"</span>"+
+         "<span id='corrcont' class='text' style='color:gray' maxlength='10'>" +"-"+ e.contenido+"</span>"+"<span id='trash' class='glyphicon glyphicon-trash' style='float:right' onClick='LOGIN.eliminarcorreoenviados("+i+");'></span>"+"<nav><div id="+"E"+i+"></div></nav>"+
+         "<div id="+"M"+i+" class='ocultar'class='pr colocar'><div ><div class= 'colornuevo class='animated fadeIn sombra2'>"+
+         "<header class=he><div><p class=txt izq>"+e.asunto+"</p><div class=>"+
+         "<a title=Eliminar correo>"+
+         "<img class='padding' title='Responder mensaje' src='Imagenes/responder.png'><a/>"+
+         "<a onclick='document.getElementById(culto).style.display = 'none';' title=Eliminar correo><img class='padding' title='Editar mensaje' src='Imagenes/edit.png'>"+
+         "<a/><a href='eliminar.html' title='Eliminar correo'><img class='padding' src='Imagenes/trash.png'><a/>"
+         +"</div></div></header><hr>"+ "<div class='div'><p class='txt-izq'>"+e.para+"</p>"+"<div>"+"<header>" + e.asunto +"</header>"+"</div><div>"+"<p class='contenido'>"+e.contenido+"</p>"+"</div>"+"</div>"
+         +"</div></div></div>"+"</div>"
+       }else{
+        user_html = user_html + "<nav class='bandeja'>"+"<p class='hoy salto'>"+"Aca se mostraran tus mensajes que se encuentran en borrador"+"</p>"+"</nav>";
+      }
     }
     $('#correos_borrados').html(user_html);
 
@@ -271,11 +273,11 @@ $('#ienviados').html(contador);
     if(i===aidi){
       var maniacs="C"+i;
       var mos=salida[i];
-      borrador_html = borrador_html +"<div id="+"M"+i+" class='ocultar pr colocar animated fadeOutDown'><div ><div class=colornuevo sombra2>"+
-      "<header class=he><div><p class=txt izq>"+mos.asunto+"</p><div class=>"+
+      borrador_html = borrador_html +"<div id="+"M"+i+" class='ocultar pr colocar animated fadeOutDown'><div><div class=colornuevo sombra2>"+
+      "<header class='he'><div><p class=txt izq>"+mos.asunto+"</p><div class=>"+
       "<a onclick=document.getElementById('light').style.display='block';document.getElementById('fade').style.display='block' title=Eliminar correo>"+
       "<img class='padding' title='Responder mensaje' src='Imagenes/responder.png'><a/>"+
-      "<a href='editar.html' title=Eliminar correo><img class='padding' title='Editar mensaje' src='Imagenes/edit.png'>"+
+      "<a onClick='' title=Eliminar correo><img class='padding' title='Editar mensaje' src='Imagenes/edit.png'>"+
       "<a/><a href='eliminar.html' title='Eliminar correo'><img class='padding' src='Imagenes/trash.png'><a/>"
       +"</div></div></header><hr>"+ "<div class='div'><p class='txt-izq'>"+mos.para+"</p>"+"<div>"+"<header>" + mos.asunto +"</header>"+"</div><div>"+"<p class='contenido'>"+mos.contenido+"</p>"+"</div>"+"</div>"
       +"</div></div></div>";
@@ -292,7 +294,7 @@ $('#ienviados').html(contador);
     if(i===aid){
       var mani="E"+i;
       var envi=enviado[i];
-      enviado_html = enviado_html +"<div id="+"M"+i+" class='ocultar pr colocar'><div ><div class=colornuevo sombra2>"+
+      enviado_html = enviado_html +"<div id="+"M"+i+" class='ocultar pr colocar animated fadeOutDown'><div><div class=colornuevo sombra2>"+
       "<header class=he><div><p class=txt izq>"+envi.asunto+"</p><div class=>"+
       "<a onclick=document.getElementById('light').style.display='block';document.getElementById('fade').style.display='block' title=Eliminar correo>"+
       "<img class='padding' title='Responder mensaje' src='Imagenes/responder.png'><a/>"+
@@ -308,65 +310,115 @@ $('#ienviados').html(contador);
   debugger;
 
   var culto="M"+aidi2;
-  document.getElementById(culto).style.display = 'block';
+  if(conti===0){
+    conti=conti+1;
+    document.getElementById(culto).style.display = 'block';
   //document.getElementById(culto).style.display = 'none';
   //document.getElementById('corrcont').style.display = 'none';
   //document.getElementById('corrasun').style.display = 'none';
-  
+}else{
+  conti=0;
+  document.getElementById(culto).style.display = 'none';
+}
 
 
+},contoculto2:function(aidi4){
+  debugger;
+  var cul="O"+aidi4;
+  if(conta===0){
+    conta=conta+1;
+    document.getElementById(cul).style.display = 'block';
+  //document.getElementById(culto).style.display = 'none';
+  //document.getElementById('corrcont').style.display = 'none';
+  //document.getElementById('corrasun').style.display = 'none';
+}else{
+  conta=0;
+  document.getElementById(cul).style.display = 'none';
+}
 },eliminarcorreoborrador:function(ec){
-   debugger;
-  var bor=new Array();
-  this.ec=ec;
-  borrarbo= JSON.parse(localStorage.getItem("Borradores"));
-  for (var i = 0; i < borrarbo.length; i++) {
-     if(i===ec&&borrarbo.length===1){
-      localStorage.removeItem("Borradores");
-     }
-      if(i==ec){
+ debugger;
+ var bor=new Array();
+ this.ec=ec;
+ borrarbo= JSON.parse(localStorage.getItem("Borradores"));
+ localStorage.removeItem("Borradores");
+ for (var i = 0; i < borrarbo.length; i++) {
+   if(i===ec&&borrarbo.length===1){
+    localStorage.removeItem("Borradores");
+  }
+  if(i==ec){
 
   }else{
     var bora={'User':borrarbo[i].User,'para':borrarbo[i].para,'asunto':borrarbo[i].asunto,'contenido':borrarbo[i].contenido};
-      bor.push(bora);
-      localStorage.removeItem("Borradores");
-     localStorage.setItem("Borradores",JSON.stringify(bor));
-   }
-     }
+    bor.push(bora);
+    localStorage.setItem("Borradores",JSON.stringify(bor));
+  }
+}
 
-     LOGIN.cargarcorreos();
+LOGIN.cargarcorreos();
+
 },eliminarcorreoenviados:function(ec){
-   debugger;
-  var bor=new Array();
-  this.ec=ec;
-  borrarbo= JSON.parse(localStorage.getItem("Enviados"));
-  for (var i = 0; i < borrarbo.length; i++) {
-     if(i===ec&&borrarbo.length===1){
-      localStorage.removeItem("Enviados");
-     }
-      if(i==ec){
+ debugger;
+ var bor=new Array();
+ this.ec=ec;
+ borrarbo= JSON.parse(localStorage.getItem("Enviados"));
+ for (var i = 0; i < borrarbo.length; i++) {
+   if(i===ec&&borrarbo.length===1){
+    localStorage.removeItem("Enviados");
+  }
+  if(i==ec){
 
   }else{
     var bora={'User':borrarbo[i].User,'para':borrarbo[i].para,'asunto':borrarbo[i].asunto,'contenido':borrarbo[i].contenido};
-      bor.push(bora);
-      localStorage.removeItem("Enviados");
-     localStorage.setItem("Enviados",JSON.stringify(bor));
-   }
-     }
+    bor.push(bora);
+    localStorage.removeItem("Enviados");
+    localStorage.setItem("Enviados",JSON.stringify(bor));
+  }
+}
 
-     LOGIN.cargarcorreos();
+LOGIN.cargarcorreos();
 
 },cargarnombre:function(){
-   var uuser = JSON.parse(localStorage.getItem("Online")).online;
+ var uuser = JSON.parse(localStorage.getItem("Online")).online;
  $('#nomuser').html(uuser);
 
 },editarcorreoborrador:function(edit){
-  this.edit=edit;
+  debugger;
+  this.edit = edit;
+  document.getElementById("M"+edit).style.display='none';
   editarcorreo= JSON.parse(localStorage.getItem("Borradores"));
   for (var i = 0; i < editarcorreo.length; i++) {
-    
-  };
+    if(edit===i){
+      var editarmensaje= "<div class='espacio top'>"+
+      "<div id="+"O"+i+" onClick='LOGIN.contoculto2("+edit+");'></div><div class='colornuevo sombra2'><header class='he'><div><input id='ediasunto' class='ta color' type=text value="+editarcorreo[i].asunto+"></div>"+
+      "</header><hr><div class='div'><input id='edipara' class='ta color' type=text value="+editarcorreo[i].para+">"+
+      "<div></br><textarea id='edicontenido' class='tamano3 col' type=text placeholder=Escribe aqui>"+editarcorreo[i].contenido+"</textarea>"+
+      "<button class='boton b2 bt' type=button onClick='LOGIN.guardaredicion("+i+");'><span class='glyphicon glyphicon-ok'></span></button><button class='boton b2 bt colo sp' type=button><span class='glyphicon glyphicon-remove'></span></button></div></br></div></div></div>";
+      break;
+    }
 
+  }
+  document.getElementById("C"+edit).innerHTML=editarmensaje;
+
+},guardaredicion:function(ed){
+  debugger;
+ var eddit=new Array();
+ this.ed=ed;
+ var edita;
+ var edpara = document.getElementById("edipara").value;
+ var edasunto = document.getElementById("ediasunto").value;
+ var edcontenido = document.getElementById("edicontenido").value;
+ ecorreo = JSON.parse(localStorage.getItem("Borradores"));
+ localStorage.removeItem("Borradores");
+ for (var i = 0; i < ecorreo.length; i++) {
+  if(i===ed&&ecorreo.length===1){
+    edita={'User':ecorreo[i].User,'para':edpara,'asunto':edasunto,'contenido':edcontenido};
+  }if(i===ed){
+    edita={'User':ecorreo[i].User,'para':edpara,'asunto':edasunto,'contenido':edcontenido};
+  }else{
+    edita={'User':ecorreo[i].User,'para':ecorreo[i].para,'asunto':ecorreo[i].asunto,'contenido':ecorreo[i].contenido};
+    }
+    eddit.push(edita);
+    localStorage.setItem("Borradores",JSON.stringify(eddit)); 
 }
-
+}
 };
